Validate and transform calculated value
